[common.fixed_map] fixed several bugs
The const variant of find() did never compile bc of the missing *this. In
addition the getIdx(key) would have returned -1l for a missing key while
the ctor of the const_iterator expects a std::size_t. To resolve this
situation I changed getIdx to return a std::size_t and return N for a
missing key which is in line with the overall syntax in this class
concerning end() iterators.