[logging] refactor FileBuffer
Previously FileBuffer was actually buffering to file _and_ `cout`
at the same time.
This is behaviour is now split, with common functionality
in OstreamBuffer. The old behviour is kept by using
a new DualLogStream in the actual logger. This uses the
extensible CombinedBuffer that can be used as a meta container
for any SuspendableStrBuffer.