[cmake] rewire format target
`add_format` now only takes one argument: a directory that is used
the base dir for recursive globbing files into clang-format
Usages of the macros are adjusted accordingly. It's still possible to
use it multiple times in a project, if one for whatever reason
does not wnat to throw it at the source root. Only one `format`
target will be generated either way, with one subtarget per
`add_format` invocation.